Replace Assert.ReferenceEquals call with Assert.AreSame
The intention must have been to assert that the two references
are identical (refering to the same object).

Calling Assert.ReferenceEquals will not have the desired effect
since it will use System.Object.ReferenceEquals; not throwing an
exception when the two objects does not match.

Of the total 29 occurances of Assert.ReferenceEquals, two tests
failed after being replaced with Assert.AreSame.
I have modified the test case to assert what I belive was the
intention of the author.