Show regression with protected constructor check
Add tests to exercise construction and checking resolution of classes that have
only protected constructors. A new test that asserts that CanResolve of a class with only a protected constructor will return false will fail.  The regression
was introduced in
https://github.com/grumpydev/TinyIoC/commit/3891ab9428acd6bd8b4c29f6ead7808a50f423ca.
The original use of type.GetConstructors() only returend public
constructors - no protected ones.
Resolution of a protected constructor must throw and checking for resolution must also fail.  On the other hand, resolving a concrete subclass of such a class should succeed if the subclass has a public constructor that has its dependencies met.